Abstracted layer menu into it's own component
